+++ dashboard/src/scenes/search/index.js
@@ -15,9 +15,9 @@ import Search from '../../components/search';
 import TagTeam from '../../components/TagTeam';
 import { organisationState, teamsState } from '../../recoil/auth';
 import { actionsState, CANCEL, DONE, sortActionsOrConsultations } from '../../recoil/actions';
-import { personsState } from '../../recoil/persons';
+import { personsState, sortPersons } from '../../recoil/persons';
 import { relsPersonPlaceState } from '../../recoil/relPersonPlace';
-import { territoriesState } from '../../recoil/territory';
+import { sortTerritories, territoriesState } from '../../recoil/territory';
 import { selector, selectorFamily, useRecoilValue } from 'recoil';
 import { itemsGroupedByPersonSelector, onlyFilledObservationsTerritories, personsObjectSelector } from '../../recoil/selectors';
 import PersonName from '../../components/PersonName';
@@ -230,12 +230,14 @@ const personsWithFormattedBirthDateSelector = selector({
 const personsFilteredBySearchForSearchSelector = selectorFamily({
   key: 'personsFilteredBySearchForSearchSelector',
   get:
-    ({ search }) =>
+    ({ search, sortBy, sortOrder }) =>
     ({ get }) => {
       const persons = get(personsWithFormattedBirthDateSelector);
       const personsPopulated = get(itemsGroupedByPersonSelector);
       if (!search?.length) return [];
-      return filterBySearch(search, persons).map((p) => personsPopulated[p._id]);
+      return filterBySearch(search, persons)
+        .map((p) => personsPopulated[p._id])
+        .sort(sortPersons(sortBy, sortOrder));
     },
 });
 
@@ -244,7 +246,9 @@ const Persons = ({ search, onUpdateResults }) => {
   const teams = useRecoilValue(teamsState);
   const organisation = useRecoilValue(organisationState);
 
-  const data = useRecoilValue(personsFilteredBySearchForSearchSelector({ search }));
+  const [sortBy, setSortBy] = useLocalStorage('person-sortBy', 'name');
+  const [sortOrder, setSortOrder] = useLocalStorage('person-sortOrder', 'ASC');
+  const data = useRecoilValue(personsFilteredBySearchForSearchSelector({ search, sortBy, sortOrder }));
 
   useEffect(() => {
     onUpdateResults(data.length);
@@ -276,6 +280,10 @@ const Persons = ({ search, onUpdateResults }) => {
             {
               title: '',
               dataKey: 'group',
+              onSortOrder: setSortOrder,
+              onSortBy: setSortBy,
+              sortOrder,
+              sortBy,
               small: true,
               render: (person) => {
                 if (!person.group) return null;
@@ -288,10 +296,21 @@ const Persons = ({ search, onUpdateResults }) => {
                 );
               },
             },
-            { title: 'Nom', dataKey: 'name' },
+            {
+              title: 'Nom',
+              dataKey: 'name',
+              onSortOrder: setSortOrder,
+              onSortBy: setSortBy,
+              sortOrder,
+              sortBy,
+            },
             {
               title: 'Vigilance',
               dataKey: 'alertness',
+              onSortOrder: setSortOrder,
+              onSortBy: setSortBy,
+              sortOrder,
+              sortBy,
               render: (p) =>
                 p.alertness ? (
                   <ExclamationMarkButton
@@ -301,7 +320,15 @@ const Persons = ({ search, onUpdateResults }) => {
                 ) : null,
             },
             { title: 'Équipe(s) en charge', dataKey: 'assignedTeams', render: (person) => <Teams teams={teams} person={person} /> },
-            { title: 'Suivi(e) depuis le', dataKey: 'followedSince', render: (p) => formatDateWithFullMonth(p.followedSince || p.createdAt || '') },
+            {
+              title: 'Suivi(e) depuis le',
+              dataKey: 'followedSince',
+              onSortOrder: setSortOrder,
+              onSortBy: setSortBy,
+              sortOrder,
+              sortBy,
+              render: (p) => formatDateWithFullMonth(p.followedSince || p.createdAt || ''),
+            },
           ].filter((c) => organisation.groupsEnabled || c.dataKey !== 'group')}
         />
       </StyledBox>
@@ -481,11 +508,13 @@ const Comments = ({ search, onUpdateResults }) => {
 const Territories = ({ search, onUpdateResults }) => {
   const history = useHistory();
   const territories = useRecoilValue(territoriesState);
+  const [sortBy, setSortBy] = useLocalStorage('territory-sortBy', 'name');
+  const [sortOrder, setSortOrder] = useLocalStorage('territory-sortOrder', 'ASC');
 
   const data = useMemo(() => {
     if (!search?.length) return [];
-    return filterBySearch(search, territories);
-  }, [search, territories]);
+    return filterBySearch(search, territories).sort(sortTerritories(sortBy, sortOrder));
+  }, [search, territories, sortBy, sortOrder]);
 
   useEffect(() => {
     onUpdateResults(data.length);
@@ -506,10 +535,40 @@ const Territories = ({ search, onUpdateResults }) => {
           onRowClick={(territory) => history.push(`/territory/${territory._id}`)}
           rowKey="_id"
           columns={[
-            { title: 'Nom', dataKey: 'name' },
-            { title: 'Types', dataKey: 'types', render: ({ types }) => (types ? types.join(', ') : '') },
-            { title: 'Périmètre', dataKey: 'perimeter' },
-            { title: 'Créé le', dataKey: 'createdAt', render: (territory) => formatDateWithFullMonth(territory.createdAt || '') },
+            {
+              title: 'Nom',
+              dataKey: 'name',
+              onSortOrder: setSortOrder,
+              onSortBy: setSortBy,
+              sortOrder,
+              sortBy,
+            },
+            {
+              title: 'Types',
+              dataKey: 'types',
+              onSortOrder: setSortOrder,
+              onSortBy: setSortBy,
+              sortOrder,
+              sortBy,
+              render: ({ types }) => (types ? types.join(', ') : ''),
+            },
+            {
+              title: 'Périmètre',
+              dataKey: 'perimeter',
+              onSortOrder: setSortOrder,
+              onSortBy: setSortBy,
+              sortOrder,
+              sortBy,
+            },
+            {
+              title: 'Créé le',
+              dataKey: 'createdAt',
+              onSortOrder: setSortOrder,
+              onSortBy: setSortBy,
+              sortOrder,
+              sortBy,
+              render: (territory) => formatDateWithFullMonth(territory.createdAt || ''),
+            },
           ]}
         />
       </StyledBox>